move cooldown logic to it's own helper
