+++ litellm/proxy/proxy_config.yaml
@@ -6,6 +6,14 @@ model_list:
       vertex_project: "adroit-crow-413218"
       vertex_location: "us-central1"
       vertex_credentials: "/Users/ishaanjaffer/Downloads/adroit-crow-413218-a956eef1a2a8.json"
+  - model_name: gemini-vision
+    litellm_params:
+      model: vertex_ai/gemini-1.0-pro-vision-001
+      api_base: https://exampleopenaiendpoint-production-c715.up.railway.app/v1/projects/adroit-crow-413218/locations/us-central1/publishers/google/models/gemini-1.0-pro-vision-001
+      vertex_project: "adroit-crow-413218"
+      vertex_location: "us-central1"
+      vertex_credentials: "/Users/ishaanjaffer/Downloads/adroit-crow-413218-a956eef1a2a8.json"
+
   - model_name: fake-azure-endpoint
     litellm_params:
       model: openai/429
@@ -14,9 +22,7 @@ model_list:
 
 general_settings: 
  master_key: sk-1234 
- alerting: ["slack"]
- alerting_threshold: 0.00001
 
 litellm_settings:
-  callbacks: ["otel"]
-
+  callbacks: ["prometheus"]
+  allowed_fails: 10000
